refactor account file into multiple smaller files
